feat: Rename generator
